+++ frontend/src/components/Table/Table.tsx
@@ -9,23 +9,6 @@ import './Table.css';
 
 type TableProps = {};
 export function Table({}: TableProps) {
-  const { subscribeToEvent } = useSocket();
-  // TODO: push down to appropriate component
-  useEffect(() => {
-    const subscribeToPlayerJoined = subscribeToEvent('player_joined', () => {
-      console.log('Player sat down');
-    });
-
-    const subscribeToPlayerLeave = subscribeToEvent('player_left', () => {
-      console.log('Player left the table');
-    });
-
-    return () => {
-      subscribeToPlayerJoined();
-      subscribeToPlayerLeave();
-    };
-  }, [subscribeToEvent]);
-
   return (
     <>
       <div id="table">
