+++ frontend/src/hooks/useSocket.ts
@@ -15,22 +15,16 @@ export function useSocket() {
       console.log('Hello from server');
     };
 
-    const onGameReady = () => {
-      console.log('Starting Game');
-    };
-
     // Register event handlers
     socket.on('connect', onConnect);
     socket.on('disconnect', onDisconnect);
     socket.on('hello_from_server', onHelloFromServer);
-    socket.on('start_game', onGameReady);
 
     // Cleanup on component unmount
     return () => {
       socket.off('connect', onConnect);
       socket.off('disconnect', onDisconnect);
       socket.off('hello_from_server', onHelloFromServer);
-      socket.off('start_game', onGameReady);
     };
   }, []);
